LayoutTests/fast/selectors/querySelector-pseudo-element-inside-functional-pseudo-class-not.html

<!doctype html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div style="display:none" id="test">
<div>
    <p class="ng" id="target">Cocoa and Cappuccino make us happy.</p>
</div>
</div>
</body>
<script>
description('This test makes sure that querySelector with pseudo-element inside functional pseudo class :not is syntax error.');
shouldThrow('document.querySelectorAll("#test div :not(:first-line)")');
</script>
<script src="../../resources/js-test-post.js"></script>
</html>
