LayoutTests/fast/selectors/querySelector-pseudo-element-inside-functional-pseudo-class-not-expected.txt

This test makes sure that querySelector with pseudo-element inside functional pseudo class :not is syntax error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("#test div :not(:first-line)") threw exception SyntaxError: '#test div :not(:first-line)' is not a valid selector..
PASS successfullyParsed is true

TEST COMPLETE
